+++ src/eps/tutorials/ex2.c
@@ -148,4 +148,9 @@ int main(int argc,char **argv)
          args: -eps_type feast -eps_interval 1.1,1.25 -eps_ncv 64
          requires: feast
 
+   test:
+      suffix: 3
+      args: -n 30 -m 30 -eps_interval 3.9,4.15 -st_type sinvert -st_pc_type cholesky -terse
+      requires: !single
+
 TEST*/